requirements.txt
backtrader
backtrader[plotting]
yfinance
matplotlib
ipywidgets==7.7.1
python-binance
python-dotenv
kafka
pydantic
tabulate
confluent-kafka
rich
fastapi
psycopg2-binary
SQLAlchemy
alembic
pytest
passlib